refactor: layout directory
